lostfound/settings.py
from pathlib import Path
import os
from dotenv import load_dotenv
load_dotenv()

BASE_DIR = Path(__file__).resolve().parent.parent
SECRET_KEY = os.getenv("SECRET_KEY", "dev-secret")
DEBUG = True
ALLOWED_HOSTS = ["127.0.0.1", "localhost"]

DATABASES = {
    "default": {
        "ENGINE": "djongo",
        "NAME": "lostfound",
        "ENFORCE_SCHEMA": False,
        "CLIENT": {"host": os.getenv("MONGO_URI", "mongodb://localhost:27017")},
    }
}

LOGIN_URL = "/login"
LOGIN_REDIRECT_URL = "/dashboard"
LOGOUT_REDIRECT_URL = "/login"