requirements.txt
Django>=5.1,<5.2
djongo>=1.3.6
pymongo>=4.7.2
python-dotenv>=1.0.1